Refactor: rename to highlight meaning of folders.
