+++ rework_experiments/lib/navigation_experiment/example_with_hero.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:rework_experiments/navigation/lib/navigation/internal/wolt_modal_sheet_page.dart';
-import 'package:rework_experiments/navigation/lib/navigation/internal/wolt_modal_sheet_path.dart';
-import 'package:rework_experiments/navigation/lib/navigation/internal/wolt_modal_sheet_path_settings.dart';
-import 'package:rework_experiments/navigation/lib/navigation/internal/wolt_modal_sheet_navigator.dart';
-import 'package:rework_experiments/navigation/lib/navigation/external/wolt_modal_sheet_route.dart';
-import 'package:rework_experiments/wolt_page_layout/wolt_page_layout.dart';
+import 'package:rework_experiments/navigation_experiment/lib/navigation/internal/wolt_modal_sheet_page.dart';
+import 'package:rework_experiments/navigation_experiment/lib/navigation/internal/wolt_modal_sheet_path.dart';
+import 'package:rework_experiments/navigation_experiment/lib/navigation/internal/wolt_modal_sheet_path_settings.dart';
+import 'package:rework_experiments/navigation_experiment/lib/navigation/internal/wolt_modal_sheet_navigator.dart';
+import 'package:rework_experiments/navigation_experiment/lib/navigation/external/wolt_modal_sheet_route.dart';
+import 'package:rework_experiments/wolt_page_layout_experiment/wolt_page_layout.dart';
 
 void main() {
   runApp(const MyApp());
