+++ rework_experiments/lib/navigation_experiment/lib/navigation/external/wolt_modal_sheet_route.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:rework_experiments/navigation/lib/wolt_modal_sheet.dart';
-import 'package:rework_experiments/navigation/lib/navigation/internal/wolt_modal_sheet_path.dart';
-import 'package:rework_experiments/navigation/lib/navigation/internal/wolt_modal_sheet_path_settings.dart';
+import 'package:rework_experiments/navigation_experiment/lib/wolt_modal_sheet.dart';
+import 'package:rework_experiments/navigation_experiment/lib/navigation/internal/wolt_modal_sheet_path.dart';
+import 'package:rework_experiments/navigation_experiment/lib/navigation/internal/wolt_modal_sheet_path_settings.dart';
 
 /// An imperative way to show [WoltModalSheet].
 Future<T?> showWoltModalSheet<T>({
