+++ rework_experiments/lib/navigation_experiment/lib/navigation/navigation_size_adapter.dart
